Support dynamic default values
When changing default values in settings.py it is sometimes desirable
for those changes to take effect for any unset config values.

As it currently works the default value is persisted to storage
when accessed.

This commit introduces a new setting:

CONSTANCE_DYNAMIC_DEFAULTS=False

That when set to True simply avoids saving the default value
to the storage backend.