Add caching redis backend
Quite often configuration is a rarely updated. And get current value in time is not so important. But redis backend always retrieves a current value.
Due to this simple access to the config has a sudden overhead especially in cycles. Programmer should always remeber about it and cache values manually.
Even a small timeout can drastically increase performance.